feat:add new pages
