+++ index.js
@@ -2,9 +2,11 @@ const env = process.env.NODE_ENV || 'development';
 const mongoose = require('mongoose')
 const config = require('./config/config')[env];
 
-
 const express = require('express');
 const indexRouter = require('./routes')
+const authRouter= require('./routes/auth.js')
+const cubeRouter=require('./routes/cube.js')
+const accessoryRouter=require('./routes/accessory.js')
 const app = express();
 
 
@@ -19,6 +21,11 @@ mongoose.connect(url, {
   .catch(err => console.log(err))
 
 require('./config/express')(app);
+
+app.use('/', authRouter);
+app.use('/', cubeRouter);
+app.use('/', accessoryRouter);
 app.use('/', indexRouter)
+
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
