+++ views/details.hbs
@@ -5,6 +5,8 @@
     <p><span>Difficulty level:</span>{{difficulty}}</p>
     <a class="btn" href="/">Back</a>
     <a class="btn" href="/attach/accessory/{{_id}}">Attach accessory</a>
+    <a class="btn" href="/{{id}}">Delete</a>
+
 
 </div>
 <h2>Accessories</h2>
