+++ views/partials/header.hbs
@@ -5,5 +5,10 @@
     <li><a href="/create">Add a Cube</a></li>
      <li><a href="/create/accessory">Create Accessory</a></li>
     <li><a href="/about">About</a></li>
+
+
+     <li><a href="/login">Login</a></li>
+    <li><a href="/signup">Register</a></li>
+    <li><a href="/">Logout</a></li>
   </ul>
 </header>
